Fixed button command, working on counting command
buttonPrompt is now a working example of a two-phase prompt (three counting original invocation), implemented with both collectors and awaitMessages. Next is to get the countTo to work right, as an example of a command with an arbitrary number of follow-up messages, which will be crucial for making a trivia game.